Fix the disasterous performance
Disables xdebug on the CLI to make tests run in a few seconds rather than taking minutes.

Also adds support for using docker-sync for "native" performance in the docker container. If you choose to not use it, vagrant will mount the files using the standard osx method which is slower, but still within reasonable bounds.